+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to CQFill
 
+### 0.2.0 (April 26, 2021)
+
+- Adds a PostCSS plugin.
+- Prevent script from throwing in non-DOM environments.
+
 ### 0.1.1 (April 25, 2021)
 
 - Fixes the IIFE export.